Decouple from shared credentials
Committing application configuration such as this instead of relying on
environment variables means that every environment uses the same secret
key base.

This is dangerous in the case of Beacon, as the secret_key_base is used
to encrypt and decrypt the relational data for models.

By shifting to using SECRET_KEY_BASE from within the environemnt,
and removing config/credentials.yml.enc we ensure that each deployment
may manage it's own encryption keys.